enhance switcher
