+++ src/SmartTalk.Core/Services/Audio/IAudioModelProvider.cs
@@ -1,0 +1,12 @@
+using SmartTalk.Core.Ioc;
+using SmartTalk.Messages.Commands.SpeechMatics;
+using SmartTalk.Messages.Enums.Audio;
+
+namespace SmartTalk.Core.Services.Audio;
+
+public interface IAudioModelProvider : IScopedDependency
+{
+    public AudioModelProviderType ModelProviderType { get; set; }
+    
+    Task<string> ExtractAudioDataFromModelProviderAsync(AnalyzeAudioCommand command, CancellationToken cancellationToken);
+}
