+++ src/SmartTalk.Messages/Enums/Audio/AudioModelProviderType.cs
@@ -1,0 +1,7 @@
+namespace SmartTalk.Messages.Enums.Audio;
+
+public enum AudioModelProviderType
+{
+    Google,
+    OpenAi
+}
